% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_contamination_score.R
\name{ARCScore}
\alias{ARCScore}
\title{Calculate the ambient RNA contamination (ARC) score}
\usage{
ARCScore(count_mat, background_idx)
}
\arguments{
\item{count_mat}{(matrix of num) Input gene-by-barcode count matrix.
Ideally this should be the raw count matrix with all genes and barcodes
without any filtering.
Can be either standard matrix format or sparse matrix format.}

\item{background_idx}{(vector of int) Indices of background barcodes in
\code{count_mat}. For spatial data, these are background spots not
covered by tissue. For single-cell data, these are empty droplets not
containing cells.}
}
\value{
(num) The ARC score of given data.
}
\description{
The ARC score is intended for measuring the level of
contamination caused by ambient RNAs for a given sptial transcriptomics
or droplet-based single-cell RNA-seq data. Intuitively, this score is a lower
bound of the average proportion of contaminated expressions in tissue
spots or cell droplets.
}
\examples{

data(mbrain_raw)
data(mbrain_slide_info)
mbrain_raw <- mbrain_raw[,mbrain_slide_info$slide$barcode]
background_index <- which(mbrain_slide_info$slide$tissue==0)
ARCScore(mbrain_raw, background_index)
}
