% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbrain_slide_info.R
\docType{data}
\name{mbrain_slide_info}
\alias{mbrain_slide_info}
\title{Example 10x Visium spatial data: slide information}
\format{
A list of slide information containing an object of class
\code{"data.frame"} and an object of class \code{"grob"}.
}
\source{
\href{https://support.10xgenomics.com/spatial-gene-expression/datasets/1.0.0/V1_Adult_Mouse_Brain}{V1_Adult_Mouse_Brain}
}
\usage{
data(mbrain_slide_info)
}
\description{
This dataset contains one list \code{mbrain_slide_info}.
The first slot \code{slide} of \code{mbrain_slide_info} is a
data.frame of the spatial information for spots. The second slot \code{grob}
is a Grob object of the tissue image. The original dataset can be found at
https://support.10xgenomics.com/spatial-gene-expression/datasets/1.0.0/V1_Adult_Mouse_Brain.

If users read raw data using \code{read10xSlide}, they should get
exactly the same format as the objects in this dataset.
}
\examples{
data(mbrain_slide_info)
str(mbrain_slide_info)

}
\keyword{datasets}
