% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpotClean-package.R
\docType{package}
\name{SpotClean-package}
\alias{SpotClean}
\alias{SpotClean-package}
\title{SpotClean: a computational method to adjust for spot swapping in
spatial transcriptomics data}
\description{
SpotClean is a computational method to adjust for spot swapping in
spatial transcriptomics data. Recent spatial transcriptomics experiments
utilize slides containing thousands of spots with spot-specific barcodes
that bind mRNA. Ideally, unique molecular identifiers at a spot measure
spot-specific expression, but this is often not the case due to bleed from
nearby spots, an artifact we refer to as spot swapping. SpotClean is able
to estimate the contamination rate in observed data and decontaminate the
spot swapping effect, thus increase the sensitivity and precision of
downstream analyses.
}
\details{
To learn more about \code{SpotClean}, see the vignette
using \code{browseVignettes(package = "SpotClean")}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zijianni/SpotClean}
  \item Report bugs at \url{https://github.com/zijianni/SpotClean/issues}
}

}
\author{
\strong{Maintainer}: Zijian Ni \email{zni25@wisc.edu} (\href{https://orcid.org/0000-0003-1181-8337}{ORCID})

Other contributors:
\itemize{
  \item Christina Kendziorski [contributor]
}

}
