% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_heatmap.R
\name{VisualizeHeatmap}
\alias{VisualizeHeatmap}
\title{Visualize spot values on the 2D slide}
\usage{
VisualizeHeatmap(
  slide_obj,
  value,
  subset_barcodes = NULL,
  logged = TRUE,
  legend_range = NULL,
  title = "",
  legend_title = "Value"
)
}
\arguments{
\item{slide_obj}{A slide object created or inherited from
\code{CreateSlide()}.}

\item{value}{(numeric vector or chr) Either a vector of numeric values for all
spots, or the name of one gene in an assay specified in \code{assay}.
In the former case, the order of values in the vector
should match the spot barcodes in the slide object.}

\item{subset_barcodes}{(vector of chr) A subset of spot barcodes to plot.
By default it plots all spots in the slide object. This can be useful
when only plotting tissue spots or specific tissue types or regions.
Default: \code{NULL}.}

\item{logged}{(logical) Specify if the color scale is log1p transformed.
Default: \code{TRUE}.}

\item{legend_range}{(length 2 vector of num) Custom legend range of the value.
By default uses the range of the plotted values.
Default: \code{NULL}.}

\item{title}{(chr) Title of the plot. Default: \code{""}.}

\item{legend_title}{(chr) Title of the legend. Default: \code{"Value"}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Generate and visualize spot labels on the 2D slide
as a ggplot2 object. Spot labels can be given via parameter \code{label} as
external input or one column in the \code{slide} slot of the slide
object's metadata via \code{label_col}. Exactly one of \code{label}
and \code{label_col} must be specified.
}
\examples{

data(mbrain_raw)
data(mbrain_slide_info)
mbrain_obj <- CreateSlide(mbrain_raw,
                          mbrain_slide_info)
gp <- VisualizeHeatmap(mbrain_obj, "Bc1",
                     title="mbrain", legend_title="Bc1 expression")
plot(gp)
}
