% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_contamination_score.R
\name{ARCScore}
\alias{ARCScore}
\alias{ARCScore.default}
\alias{ARCScore.SummarizedExperiment}
\title{Calculate the ambient RNA contamination (ARC) score}
\usage{
ARCScore(object, ...)

\method{ARCScore}{default}(object, background_bcs, ...)

\method{ARCScore}{SummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{A gene-by-barcode count matrix or a slide object created or
inherited from \code{CreateSlide()}.
Ideally this should give the raw count matrix with all genes and barcodes
without any filtering.}

\item{...}{Arguments passed to other methods}

\item{background_bcs}{(vector of chr) Background barcodes in
\code{count_mat}. For spatial data, these are background spots not
covered by tissue. For single-cell data, these are empty droplets not
containing cells.}
}
\value{
(num) The ARC score of given data.
}
\description{
The ARC score is intended for subjectively measuring the level
of contamination caused by ambient RNAs for a given sptial transcriptomics
or droplet-based single-cell RNA-seq data. Intuitively, this score is a lower
bound of the average proportion of contaminated expressions in tissue
spots or cell droplets.

This lower bound is very conservative as it only
accounts for contamination in background spots or empty droplets, neglecting
contamination in tissue spots or cell droplets. However, it is totally
subjective, not relying on any complicated assumptions and modelling.
ARC score serves as a relative contamination measurement for comparison
among different datasets and protocols.
}
\examples{

data(mbrain_raw)
data(mbrain_slide_info)
background_bcs <- dplyr::filter(mbrain_slide_info$slide, tissue==0)$barcode
ARCScore(mbrain_raw, background_bcs)

mbrain_obj <- CreateSlide(mbrain_raw, mbrain_slide_info)
ARCScore(mbrain_obj)

}
