% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_seurat.R
\name{ConvertToSeurat}
\alias{ConvertToSeurat}
\title{Convert slide object to Seurat object}
\usage{
ConvertToSeurat(slide_obj, image_dir, slice = "slice1")
}
\arguments{
\item{slide_obj}{A slide object created or inherited from
\code{CreateSlide()}.}

\item{image_dir}{Path to directory with 10X Genomics visium image data;
should include files \code{tissue_lowres_iamge.png},
\code{scalefactors_json.json} and \code{tissue_positions_list.csv}.}

\item{slice}{Name for the stored image of the tissue slice.
Default: "slice1".}
}
\value{
A Seurat object with spatial information.
}
\description{
This function converts our slide object of class
\code{SummarizedExperiment} to Seurat object of class \code{Seurat} so that
users can directly proceed with Seurat spatial analyses pipelines.
Built based on Seurat's \code{Load10X_Spatial()}.
}
\examples{

\dontrun{
data(mbrain_raw)
data(mbrain_slide_info)
mbrain_obj <- CreateSlide(mbrain_raw,
                          mbrain_slide_info)
example_image_dir <- 'path/to/image/dir'
seurat_obj <- ConvertToSeurat(mbrain_obj, example_image_dir, "raw")
str(seurat_obj)
}

}
