% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontamination.R
\name{SpotClean}
\alias{SpotClean}
\title{Decontaminate spot swapping effect in spatial transcriptomics data}
\usage{
SpotClean(
  slide_obj,
  gene_keep = NULL,
  maxit = 30,
  tol = 1,
  candidate_radius = 5 * seq_len(6),
  verbose = TRUE
)
}
\arguments{
\item{slide_obj}{A slide object created or inherited from
\code{CreateSlide()}.}

\item{gene_keep}{(vector of chr) Gene names to keep for decontamination.
We recommend not decontaminating lowly expressed and lowly variable genes
in order to save computation time. Even if user include them, their
decontaminated expressions will not change too much from raw expressions.
When setting to \code{NULL}, \code{KeepHighGene()} will be
automatically called to filter out lowly expressed and lowly variable genes
before decontamination.
Default: \code{NULL}.}

\item{maxit}{(int) Maximum iteration for EM parameter updates. Default: 30.}

\item{tol}{(num) Tolerance to define convergence in EM parameter updates.
When the element-wise maximum difference between current and updated
parameter matrix is less than \code{tol}, parameters are considered
converged. Default: 1.}

\item{candidate_radius}{(vector of num) Candidate contamination radius.
A series of radius to try when estimating contamination parameters.
Default: {5, 10, 15, 20, 25, 30}}

\item{verbose}{(logical) Whether print progress information.
Default: \code{TRUE}}
}
\value{
A slide object where the decontaminated expression matrix is in the
"decont" assay slot and the contamination statistics are in
metadata slots. Contamination statistics include ambient RNA contamination
(ARC) score, bleeeding rate, distal rate, contamination radius,
contamination kernel weight matrix, log-likelihood value in each iteration,
estimated proportion of contamination in each tissue spot in observed data.
Since decontaminated and raw data have different columns, they can
not be stored in a single SummarizedExperiment object.
Do not overwrite the raw slide object.
}
\description{
This is the main function implementing the SpotClean method
for decontaminating spot swapping effect in spatial transcriptomics data.
}
\details{
Briefly, the contamination level for the slide is estimated based on
the total counts of all spots. UMI counts travelling around the slide are
assumed to follow Poisson distributions and modeled by a mixture of
Gaussian (proximal) and uniform (distal) kernels. The underlying
uncontaminated gene expressions are estimated by EM algorithm to
maximize the data likelihood. Detailed derivation can be found in our
manuscript.
}
\examples{

data(mbrain_raw)
data(mbrain_slide_info)
mbrain_obj <- CreateSlide(mbrain_raw,
                          mbrain_slide_info)
mbrain_decont_obj <- SpotClean(mbrain_obj, tol=10, candidate_radius=20)
mbrain_decont_obj
}
